requirements.txt
transformers>=4.11.3
datasets>=1.14.0
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.21.2
matplotlib>=3.4.3
Pillow>=8.3.1
requests>=2.26.0
accelerate==0.5.0
peft==0.3.2
timm==0.5.0